==> 00_data_collection.ipynb <==
# Get All Tweets with #PodRevDay

# Get the data
# !snscrape --jsonl twitter-hashtag "#podrevday since:2020-01-01 until:2021-11-10" > 'data/jan20_nov21.json'

import pandas as pd
from geotext import GeoText

df_tweets = pd.read_json('data/jan20_nov21.json', lines=True)

df_tweets.shape

# df_tweets.to_csv('data/jan20_oct21.json', index=False)

# Get User Data

df_tweets.user[0]

df_tweets['username'] = df_tweets['url'].str.split('/').str[3]

pod_rev_users = pd.DataFrame(df_tweets.user.to_list())

pod_rev_users.head().T

len(set(pod_rev_users.username))

pod_rev_users = pod_rev_users.drop_duplicates(subset=['username'])
len(pod_rev_users)

pod_rev_users.head().T

print(pod_rev_users.columns)
pod_rev_users.shape

# Clean User Data

keep = ['id','username', 'displayname', 'location', 'created', 'followersCount', 'friendsCount', 'url', 'verified' ]
df_users = pod_rev_users.loc[:, keep]

def location_extraction (df):
    '''Creates creates a geotext column to extract city and country info if possible'''

    df.loc[:, "location"] = df.loc[:, "location"].fillna("blank")
    df.loc[:, "geotext"] = df.loc[:, "location"].apply(GeoText)
    df.loc[:, 'city'] = df.loc[:, 'geotext'].apply(lambda x: x.cities)
    df.loc[:, 'country'] = df.loc[:, 'geotext'].apply(lambda x: x.countries)
    
    return df

df_users = location_extraction(df_users)

from geonamescache import GeonamesCache
gc = GeonamesCache()
countries = gc.get_countries()
country_info = pd.DataFrame(countries).T
country_info = country_info.set_index('geonameid').reset_index()
name_code = country_info.loc[:, ["name", "iso3"]]

us_states = ["AL", "AK", "AZ", "AR", "CA", "CO", "CT", "DC", "DE", "FL", "GA", 
             "HI", "ID", "IL", "IN", "IA", "KS", "KY", "LA", "ME", "MD", "MA",
             "MI", "MN", "MS", "MO", "MT", "NE", "NV", "NH", "NJ", "NM", "NY",
             "NC", "ND", "OH", "OK", "OR", "PA", "RI", "SC", "SD", "TN", "TX",
             "UT", "VT", "VA", "WA", "WV", "WI", "WY", "USA", "United States",
             'Seattle', "Los Angeles", "Houston", "Atlanta", "Pittsburgh"]

us_state_names = ["Alaska", "Alabama", "Arkansas", "American Samoa", "Arizona", "California", "Colorado", "Connecticut", 
               "District ", "of Columbia", "Delaware", "Florida", "Georgia", "Guam", "Hawaii", "Iowa", "Idaho", "Illinois", 
               "Indiana", "Kansas", "Kentucky", "Louisiana", "Massachusetts", "Maryland", "Maine", "Michigan", "Minnesota", 
               "Missouri", "Mississippi", "Montana", "North Carolina", "North Dakota", "Nebraska", "New Hampshire", "New Jersey", 
               "New Mexico", "Nevada", "New York", "Ohio", "Oklahoma", "Oregon", "Pennsylvania", "Puerto Rico", "Rhode Island", 
               "South Carolina", "South Dakota", "Tennessee", "Texas", "Utah", "Virginia", "Virgin Islands", "Vermont", "Washington", 
               "Wisconsin", "West Virginia", "Wyoming"]

can_prov_abbrev = {'Alberta': 'AB','British Columbia': 'BC','Manitoba': 'MB', 'New Brunswick': 'NB',
                       'Newfoundland and Labrador': 'NL', 'Northwest Territories': 'NT','Nova Scotia': 'NS','Nunavut': 'NU',
                       'Ontario': 'ON','Prince Edward Island': 'PE', 'Quebec': 'QC','Saskatchewan': 'SK','Yukon': 'YT'}

can_prov_names, can_prov_abbr = zip(*can_prov_abbrev.items())

uk = ["England", 'Wales', "Scotland", 'London', "Manchester", "Isle of Wight", "Northern Ireland", "United Kingdom", 'Bailiwick of Guernsey', "UK", "Hoxton", "Jersey"]

india_city = ["Bangalore", "Delhi", "Hyderabad", "Bengaluru"]

german_city = ["Munich", "Berlin", "eisgau","Hamburg", "Dortmund"]

south_africa = ["South Africa", "Durban", "Johannesburg"]

south_korea = ["Republic of Korea"]

uae = ['UAE', 'Dubai', 'Abu Dhabi']


def replacer(area, name):
    df_users.loc[(df_users.location.str.contains('|'.join(area))), "country"] = name    
    return df_users

def list_to_string(df):
    df["city"] = df['city'].apply(lambda x: "".join(map(str, x)))
    df["country"] = df['country'].apply(lambda x: "".join(map(str, x)))
    
    return df

def location_cleaner(df):
    df = replacer(can_prov_names, "Canada")
    df = replacer(can_prov_abbrev, "Canada")
    df = replacer(us_state_names, "United States")
    df = replacer(us_states, "United States")
    df = replacer(uk, "United Kingdom")
    df = replacer(german_city, "Germany")
    df = replacer(south_africa, "South Africa")
    df = replacer(india_city, "India")
    df = replacer(uae, 'United Arab Emirates')
    df = replacer(south_korea, "South Korea")
    df.loc[(df.location == "Italia"), "country"] = "Italy" 
    df.loc[(df.location == "Belgrade"), "country"] = "Serbia" 
    df.loc[(df.country == "PolandSerbia"), "country"] = "Poland" 
    df = list_to_string(df)
    
    return df

df_users = location_cleaner(df_users)
df_users.loc[(df_users.country == "PolandSerbia"), "country"] = "Poland" 

df_users.country.value_counts()

df_users_full = pd.merge(left=df_users,
                    right=name_code,
                    how='left',
                    left_on='country',
                    right_on='name')

df_users_full.head()

## Save the Users

df_users.to_csv('data/user_data.csv')

# Merge User's Geographic Data with Tweets

full_df = pd.merge(left=df_tweets, 
                   right=df_users, 
                   how='left',
                   on='username')


new_path = 'data/tweets_users_nov_21.csv'
full_df.to_csv(new_path, index=False)

full_df.shape

full_df.head()

import pandas as pd 

df = pd.read_csv(new_path)

df['count'] = df.groupby('username')['username'].transform('count')

df.sort_values(by='count')

df.columns

keep = ['username',  'date', 'id_y', 'displayname', 'location',
        'created', 'followersCount', 'friendsCount', 'url_y', 'verified',
        'city', 'country', 'count']

df.loc[:, keep].drop_duplicates().sort_values(by='count').to_csv('data/user_data_count_11_21.csv', index=False)
