==> 02_categorizer.ipynb <==
import httpx
import pandas as pd
import requests 

from bs4 import BeautifulSoup


display.max_colwidth = 100

df_full = pd.read_csv('data/tweets_users_sept_21.csv', parse_dates=['date'])

df_full.info()

def cleaner(text):
    return text.replace("['", "").replace("']", '').replace("[]", "")

df_full['outlinks_cleaned'] = df_full["outlinks"].apply(lambda x: cleaner(x))

df_full["outlinks_cleaned"].value_counts()

podchaser_reviews = df_full[df_full["outlinks_cleaned"].str.startswith("https://www.podchaser")]
len(podchaser_reviews)

pd.set_option('display.max_colwidth', 100)

podchaser_reviews.loc[:5, ['outlinks', 'outlinks_cleaned']]

def splitter(text):
    return "/".join(text.split('/', 5)[:5])

podchaser_reviews['podcast_url'] = podchaser_reviews["outlinks_cleaned"].apply(lambda x: splitter(x))

podchaser_reviews.tail()

len(podchaser_reviews[podchaser_reviews['outlinks_cleaned'].str.contains("podcasts/0/")])


easy_podchaser = podchaser_reviews[~podchaser_reviews['outlinks_cleaned'].str.contains("podcasts/0/")]

easy_podchaser.head()

test = easy_podchaser.iloc[0, -1]
test

url = test
page = httpx.get(url)

page

# headers = {'User-Agent': 'Mozilla/5.0 (X11; Linux x86_64) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/92.0.4515.131 Safari/537.36',
#             'Referer': 'https://www.google.com'
# }
# page = requests.get(url, headers=headers)
soup = BeautifulSoup(page.content, 'html.parser')

results = soup.find(id='app')
soup.find('a', class_="[object Object]").get_text()

soup.find('div', class_='_nymwu9').get_text()

test2 = easy_podchaser.iloc[1, -1]
test2

def get_description(url):
    page = httpx.get(url)
    soup = BeautifulSoup(page.content, 'html.parser')
    description = soup.find('div', class_='_nymwu9').get_text()

    return description

get_description(test2)

test3 = easy_podchaser.iloc[-1, -1]
test3

get_description(test3)

def get_description(url):
    page = httpx.get(url)
    soup = BeautifulSoup(page.content, 'html.parser')
    description = soup.find('div', class_='_nymwu9').get_text()

    return description

page3 = httpx.get(test3)

soup = BeautifulSoup(page3.content, 'html.parser')

description = soup.findAll('a', class_="[object Object]")
title = soup.find('h1', class_='_tn1787').get_text() 

soup.find(id="page-about-section").get_text()


title

[d.get_text() for d in description]

def get_categories(url):
    page = httpx.get(url)
    soup = BeautifulSoup(page.content, 'html.parser')
    description = soup.findAll('a', class_="[object Object]")
    categories = [d.get_text() for d in description]
    # title = soup.find('h1', class_='_tn1787').get_text() 

    return categories

def get_titles(url):
    try: 
        page = httpx.get(url)
        soup = BeautifulSoup(page.content, 'html.parser')
        title = soup.find('h1', class_='_tn1787').get_text()
        
        return title
    except:
        return "ERROR"

get_titles(test3)

get_categories(test3)

print(title)
print(categories)

easy_podchaser['title'] = easy_podchaser['podcast_url'].apply(lambda x: get_titles(x))

easy_podchaser['categories'] = easy_podchaser['podcast_url'].apply(lambda x: get_categories(x))

easy_podchaser.to_csv('data/easy_podchaser_sept.csv', index=False)

easy_podchaser.iloc[:5, -5: ].head()

easy_podchaser[easy_podchaser['title']=='ERROR']
