==> 02_1_podchaser_categorizer.ipynb <==
import pandas as pd

from ast import literal_eval

df = pd.read_csv('data/easy_podchaser_sept.csv')
df.head()

df.shape

clean = df[df['title']!="ERROR"]

clean.shape

clean.columns

keep = ['title','podcast_url', 'categories', 'url_x']

first_try = clean.loc[:, keep]

first_try.info()

first_try['categories'] = first_try['categories'].apply(literal_eval)

podchaser_categories = first_try.explode('categories')

print(podchaser_categories['categories'].nunique())
print(podchaser_categories['categories'].unique())

# pd.set_option('display.max_columns', None)
# pd.set_option('display.expand_frame_repr', False)
pd.set_option('max_colwidth', None)

podchaser_categories[podchaser_categories['categories'].isna()][~]

podchaser_categories.title.nunique()

podchaser_categories.shape

clean = podchaser_categories[~podchaser_categories['categories'].isna()]
clean.title.nunique()

clean.to_csv('data/podcasts_categories_08_2021.csv', index=False)
