==> 03_semantic_analysis.ipynb <==
import pandas as pd
import re
import spacy

from spacytextblob.spacytextblob import SpacyTextBlob
from spacymoji import  Emoji

display.max_colwidth = 100

RANDOM_STATE = 42

nlp = spacy.load('en_core_web_md')

nlp.add_pipe('emoji', first=True)
nlp.add_pipe('spacytextblob')

df_full = pd.read_csv('data/tweets_users_aug_21.csv', parse_dates=['date'])

keep = ['url_x', 'date', 'content', 'renderedContent', 'id_x', 
        'user','outlinks', 'username', 'displayname', 'location',
       'followersCount', 'friendsCount']

df = df_full.loc[:, keep]

df.sample(25, random_state=RANDOM_STATE)

test = df.at[194, "content"]

test

doc = nlp(test)

doc._.polarity

doc._.assessments

a = [d.text for d in doc if not d.text.startswith('@')]
a_2 = " ".join(a)
a_2

doc2 = nlp(a_2)
print(f'polarity: {doc2._.polarity}')

# Retrieve the default token-matching regex pattern
re_token_match = spacy.tokenizer._get_regex_pattern(nlp.Defaults.token_match)

# Add #hashtag pattern
re_token_match = f"({re_token_match}|#\\w+)"

nlp.tokenizer.token_match = re.compile(re_token_match).match



doc2 = nlp(test)
doc2

b = [d.text for d in doc2 if not d.text.startswith('@')]
b = " ".join(b)
b

for i in nlp(b):
    if i._.emoji_desc:
        print(i._.emoji_desc)
